requirements.txt
joblib==1.5.2
matplotlib==3.10.6
numpy==2.2.6
pandas==2.3.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
streamlit==1.49.1
Pillow==11.3.0
opencv-python==4.12.0.88


